Update local time test for Travis CI
When this test was first written, it was designed to run locally using a time zone other than UTC. When run outside of UTC, everything works. When run using UTC, the code still works, but the test fail sbecause it is testing that they datetimes are not equal.

The new block was added to use America/New_York when the test are run using UTC, keeping the UTC test for everything else.