Add format() method
The `format()` method provides a wrapper for the `strftime()` method associated with `date`, `datetime`, and `time` objects. At this point there is no special behavior added, this method simply serves as an alternate way to get a formatted string. The main purpose for its addition is to allow for alternate behavior in the future (e.g., raising an exception when passing in a date-related directive when formatting a time).