Remove unused print statement
The `print` statement had originally been added for quick debugging during the initial writing of `ever()`. Because it occurred after the `return`, I forgot about it.

While this is a case of no harm, no foul, it prevented the Python 3.x tests from passing since `print` is a function in Python 3, not a statement.