Add methods to test for time zone aware/naive datetimes
When the `shift()` method was first given the ability to work with time zone aware datatimes, I made the assumption that checking `value.tzinfo is not None` was sufficient. After reading through the `datetime` documentation, I learned this wasn't the case. There is a second check that is required:

    A `datetime` object `d` is aware if `d.tzinfo` is not `None` and `d.tzinfo.utcoffset(d)` does not return `None`.

These new methods encapsulate these two checks.