Refactor how_many_leap_days() method
Instead of recreating all of the logic to determine if a year is a leap year and manually checking every year, this revision utilizes the calendar module and its built-in methods. The method now enforces the types of values that can be passed in, plus adds support for passing in just a year.

As this version of the method supports multiple types, the tests are being expanded to check these.